+++ src/main/java/kn18012/librarymanagement/controller/BookController.java
@@ -1,11 +1,9 @@
 package kn18012.librarymanagement.controller;
 
 import kn18012.librarymanagement.domain.Book;
-import kn18012.librarymanagement.domain.User;
 import kn18012.librarymanagement.service.AuthorService;
 import kn18012.librarymanagement.service.BookService;
 import org.springframework.data.domain.Page;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -27,13 +25,12 @@ public class BookController {
     }
 
     @GetMapping("/books-list/page/{pageNumber}")
-    public String pagedBookView(@AuthenticationPrincipal User user, @PathVariable("pageNumber") int pageNumber, @RequestParam("phrase") String phrase, Model model) {
+    public String pagedBookView(@PathVariable("pageNumber") int pageNumber, @RequestParam("phrase") String phrase, Model model) {
         // create a new page with book search results
         Page<Book> page = bookService.searchForBook(phrase, pageNumber);
         // get page content
         List<Book> books = page.getContent();
         // add attributes to template
-        model.addAttribute("user", user);
         model.addAttribute("books", books);
         model.addAttribute("phrase", phrase);
         model.addAttribute("currentPage", pageNumber);
@@ -43,8 +40,8 @@ public class BookController {
     }
 
     @GetMapping("/new-book")
-    public String addBookView(@AuthenticationPrincipal User user, Model model) {
-        model.addAttribute("user", user);
+    public String addBookView(Model model) {
+        //model.addAttribute("user", user);
         model.addAttribute("book", new Book());
         model.addAttribute("allAuthors", authorService.findAllAuthors());
         model.addAttribute("genres", bookService.findAllGenres());
@@ -52,8 +49,11 @@ public class BookController {
     }
 
     @PostMapping("/create-book")
-    public String createBook(@Valid @ModelAttribute Book book, BindingResult bindingResult) {
+    public String createBook(@Valid @ModelAttribute Book book, BindingResult bindingResult, Model model) {
+        // check for validation errors
         if(bindingResult.hasErrors()) {
+            model.addAttribute("allAuthors", authorService.findAllAuthors());
+            model.addAttribute("genres", bookService.findAllGenres());
             return "lib/new-book";
         }
 
@@ -62,12 +62,20 @@ public class BookController {
     }
 
     @GetMapping("/edit-book/{bookId}")
-    public String editBookView(@AuthenticationPrincipal User user, @PathVariable("bookId") Long id, Model model) {
-        model.addAttribute("user", user);
-        model.addAttribute("book", bookService.findById(id));
-        model.addAttribute("allAuthors", authorService.findAllAuthors());
-        model.addAttribute("genres", bookService.findAllGenres());
-        return "lib/edit-book";
+    public String editBookView(@PathVariable("bookId") Long id, Model model) {
+        //model.addAttribute("user", user);
+        // check if book exists in id
+        Book bookToEdit = bookService.findById(id);
+        if(bookToEdit == null) {
+            return "error-page";
+        } else {
+            // add atributes to model
+            model.addAttribute("book", bookToEdit);
+            model.addAttribute("allAuthors", authorService.findAllAuthors());
+            model.addAttribute("genres", bookService.findAllGenres());
+
+            return "lib/edit-book";
+        }
     }
 
     @PostMapping("/update-book/{bookId}")
@@ -76,15 +84,26 @@ public class BookController {
         if(bindingResult.hasErrors()) {
             return "lib/new-book";
         }
-
-        bookService.update(id, book);
-        return "redirect:/lib-dashboard/books-list/page/1/?phrase=&book_update=success";
+        // check if book exists in data base
+        Book bookToUpdate = bookService.findById(id);
+        if(bookToUpdate == null) {
+            return "error-page";
+        } else {
+            bookService.update(id, book);
+            return "redirect:/lib-dashboard/books-list/page/1/?phrase=&book_update=success";
+        }
     }
 
     @DeleteMapping
     @RequestMapping("/delete-book/{bookId}")
     public String deleteBook(@PathVariable("bookId") Long bookId) {
-        bookService.deleteBookById(bookId);
-        return "redirect:/lib-dashboard/books-list/page/1/?phrase=&book_delete=success";
+        // first, check if book exists in data base
+        Book bookToDelete = bookService.findById(bookId);
+        if(bookToDelete == null) {
+            return "error-page";
+        } else {
+            bookService.deleteBookById(bookId);
+            return "redirect:/lib-dashboard/books-list/page/1/?phrase=&book_delete=success";
+        }
     }
 }
